+++ examples/nextjs-with-typescript/src/pages/_app.tsx
@@ -4,7 +4,7 @@ import { AppProps } from 'next/app';
 import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { CacheProvider, EmotionCache } from '@emotion/react';
-import theme from '../../styles/theme';
+import theme from '../styles/theme';
 import createEmotionCache from '../utils/createEmotionCache';
 
 // Client-side cache, shared for the whole session of the user in the browser.
